requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
cffi==1.15.1
click==8.1.3
ffprobe-python==1.0.3
Flask==2.3.2
flask-restx==1.1.0
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
numpy==1.24.3
psycopg2-binary==2.9.6
pycparser==2.21
pydub==0.25.1
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
soundfile==0.12.1
SQLAlchemy==2.0.15
typing_extensions==4.6.1
waitress==2.1.2
Werkzeug==2.3.4
